requirements.txt
aiogram>=3.0.0
python-dotenv
aiohttp
aiofiles 